+tsql_asa dialect
