+++ src/sqlfluff/dialects/dialect_tsql.py
@@ -21,15 +21,16 @@ from sqlfluff.core.parser import (
 )
 
 from sqlfluff.core.dialects import load_raw_dialect
-from sqlfluff.dialects.tsql_keywords import RESERVED_KEYWORDS
+from sqlfluff.dialects.tsql_keywords import RESERVED_KEYWORDS, UNRESERVED_KEYWORDS
 
 ansi_dialect = load_raw_dialect("ansi")
 tsql_dialect = ansi_dialect.copy_as("tsql")
 
-# Update only RESERVED Keywords
 # Should really clear down the old keywords but some are needed by certain segments
 # tsql_dialect.sets("reserved_keywords").clear()
+# tsql_dialect.sets("unreserved_keywords").clear()
 tsql_dialect.sets("reserved_keywords").update(RESERVED_KEYWORDS)
+tsql_dialect.sets("unreserved_keywords").update(UNRESERVED_KEYWORDS)
 
 
 tsql_dialect.insert_lexer_matchers(
@@ -87,7 +88,6 @@ class StatementSegment(ansi_dialect.get_segment("StatementSegment")):  # type: i
     parse_grammar = ansi_dialect.get_segment("StatementSegment").parse_grammar.copy(
         insert=[
             Ref("CreateProcedureStatementSegment"),
-            Ref("AlterTableSwitchStatementSegment")
         ],
     )
 
@@ -355,32 +355,3 @@ class SelectStatementSegment(BaseSegment):
             Ref("OrderByClauseSegment", optional=True),
         ]
     )
-
-
-@tsql_dialect.segment()
-class AlterTableSwitchStatementSegment(BaseSegment):
-    """An `ALTER TABLE SWITCH` statement."""
-
-    type = "alter_table_statement"
-    # https://docs.microsoft.com/en-us/sql/t-sql/statements/alter-table-transact-sql?view=sql-server-ver15
-    # T-SQL's ALTER TABLE SWITCH grammar is different enough to core ALTER TABLE grammar to merit its own definition
-    match_grammar = Sequence(
-        "ALTER",
-        "TABLE",
-        Ref("SchemaNameSegment", optional=True),
-        Ref("ObjectNameSegment"),
-        "SWITCH",
-        Sequence("PARTITION", Ref("NumericLiteralSegment"), optional=True),
-        "TO",
-        Ref("SchemaNameSegment", optional=True),
-        Ref("ObjectNameSegment"),
-        Sequence(
-        "WITH",
-        #"(",
-        "TRUNCATE_TARGET",
-        #Ref("EqualsSegment"),
-        #"ON",")", 
-        optional=True),
-        Anything(),
-    )
-
