+++ src/sqlfluff/dialects/dialect_tsql_asa.py
@@ -31,9 +31,6 @@ ansi_dialect = load_raw_dialect("ansi")
 tsql_dialect = load_raw_dialect("tsql")
 tsql_asa_dialect = tsql_dialect.copy_as("tsql_asa")
 
-# Update only RESERVED Keywords
-# Should really clear down the old keywords but some are needed by certain segments
-# tsql_asa_dialect.sets("reserved_keywords").clear()
 tsql_asa_dialect.sets("unreserved_keywords").update(UNRESERVED_KEYWORDS)
 
 
@@ -129,3 +126,38 @@ class TableIndexClause(BaseSegment):
             )
         )
     )
+
+
+@tsql_asa_dialect.segment(replace=True)
+class StatementSegment(ansi_dialect.get_segment("StatementSegment")):  # type: ignore
+    """Overriding StatementSegment to allow for additional segment parsing."""
+
+    parse_grammar = tsql_dialect.get_segment("StatementSegment").parse_grammar.copy(
+        insert=[
+            Ref("AlterTableSwitchStatementSegment"),
+        ],
+    )
+
+
+@tsql_asa_dialect.segment()
+class AlterTableSwitchStatementSegment(BaseSegment):
+    """An `ALTER TABLE SWITCH` statement."""
+
+    type = "alter_table_statement"
+    # https://docs.microsoft.com/en-us/sql/t-sql/statements/alter-table-transact-sql?view=sql-server-ver15
+    # T-SQL's ALTER TABLE SWITCH grammar is different enough to core ALTER TABLE grammar to merit its own definition
+    match_grammar = Sequence(
+        "ALTER",
+        "TABLE",
+        Ref("ObjectReferenceSegment"),
+        "SWITCH",
+        Sequence("PARTITION", Ref("NumericLiteralSegment"), optional=True),
+        "TO",
+        Ref("ObjectReferenceSegment"),
+        Sequence(
+            "WITH",
+            Bracketed("TRUNCATE_TARGET", Ref("EqualsSegment"), OneOf("ON","OFF")), 
+            optional=True
+        ),
+    )
+
