+++ src/sqlfluff/dialects/tsql_asa_keywords.py
@@ -198,4 +198,5 @@ UNRESERVED_KEYWORDS = [
     "REPLICATE",
     "ROUND_ROBIN",
     "HASH",
+    "TRUNCATE_TARGET",
 ]
