+++ test/fixtures/parser/tsql_asa/alter_table_switch.sql
@@ -1,0 +1,1 @@
+ALTER TABLE [Facility].[PL_stage] SWITCH TO [Facility].[PL_BASE] WITH (TRUNCATE_TARGET = ON);
