+++ test/fixtures/parser/tsql_asa/alter_table_switch.yml
@@ -1,0 +1,29 @@
+# YML test files are auto-generated from SQL files and should not be edited by
+# hand. To help enforce this, the "hash" field in the file must match a hash
+# computed by SQLFluff when running the tests. Please run
+# `python test/generate_parse_fixture_yml.py`  to generate them after adding or
+# altering SQL files.
+_hash: 13cc73d04d16911da2f52309caf5432957a3b01d42dfd47b915c1f8e11eb4062
+file:
+  statement:
+    alter_table_statement:
+    - keyword: ALTER
+    - keyword: TABLE
+    - object_reference:
+      - identifier: '[Facility]'
+      - dot: .
+      - identifier: '[PL_stage]'
+    - keyword: SWITCH
+    - keyword: TO
+    - object_reference:
+      - identifier: '[Facility]'
+      - dot: .
+      - identifier: '[PL_BASE]'
+    - keyword: WITH
+    - bracketed:
+      - start_bracket: (
+      - keyword: TRUNCATE_TARGET
+      - comparison_operator: '='
+      - keyword: 'ON'
+      - end_bracket: )
+  statement_terminator: ;
